Split the web/auth/auth.go file
